Move websql to it's own folder
