+++ packages/datastore/datastore/src/storage/adapters/websql/WebSQLAdapter.ts
@@ -1,9 +1,9 @@
 import invariant from "tiny-invariant";
-import { StorageAdapter } from "../api/StorageAdapter";
-import { generateId } from "../LocalStorage";
-import { createLogger } from "../../utils/logger";
-import { ModelSchema } from "../../ModelSchema";
-import { Filter } from "../../filters";
+import { StorageAdapter } from "../../api/StorageAdapter";
+import { generateId } from "../../LocalStorage";
+import { createLogger } from "../../../utils/logger";
+import { ModelSchema } from "../../../ModelSchema";
+import { Filter } from "../../../filters";
 
 const logger = createLogger("sqlite");
