+++ packages/datastore/datastore/src/storage/index.ts
@@ -2,5 +2,5 @@ export * from "./LocalStorage";
 export * from "./api/CRUDEvents";
 export * from "./api/StorageAdapter";
 export * from "./api/StoreChangeEvent";
-export { WebSQLAdapter } from "./adapters/WebSQLAdapter";
+export { WebSQLAdapter } from "./adapters/websql/WebSQLAdapter";
